backend/requirement.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
celery
redis
django-cors-headers
stripe
gunicorn
django-filter
Pillow
python-dotenv